+++ cmp/grid/Grid.js
@@ -6,19 +6,21 @@
  */
 import composeRefs from '@seznam/compose-react-refs';
 import {agGrid, AgGrid} from '@xh/hoist/cmp/ag-grid';
-import {fragment, frame} from '@xh/hoist/cmp/layout';
+import {getTreeStyleClasses} from '@xh/hoist/cmp/grid';
+import {div, fragment, frame} from '@xh/hoist/cmp/layout';
 import {hoistCmp, HoistModel, useLocalModel, uses, XH} from '@xh/hoist/core';
-import {colChooser as desktopColChooser, gridFilterDialog, StoreContextMenu} from '@xh/hoist/dynamics/desktop';
+import {
+    colChooser as desktopColChooser,
+    gridFilterDialog,
+    StoreContextMenu
+} from '@xh/hoist/dynamics/desktop';
 import {colChooser as mobileColChooser} from '@xh/hoist/dynamics/mobile';
 import {convertIconToHtml, Icon} from '@xh/hoist/icon';
-import {div} from '@xh/hoist/cmp/layout';
 import {computed, observer} from '@xh/hoist/mobx';
-import {isDisplayed, logWithDebug, logDebug, apiRemoved} from '@xh/hoist/utils/js';
-import {filterConsecutiveMenuSeparators} from '@xh/hoist/utils/impl';
-import {getLayoutProps} from '@xh/hoist/utils/react';
-import {getTreeStyleClasses} from '@xh/hoist/cmp/grid';
 import {wait} from '@xh/hoist/promise';
-
+import {filterConsecutiveMenuSeparators} from '@xh/hoist/utils/impl';
+import {apiRemoved, isDisplayed, logDebug, logWithDebug} from '@xh/hoist/utils/js';
+import {getLayoutProps} from '@xh/hoist/utils/react';
 import classNames from 'classnames';
 import {
     compact,
@@ -28,9 +30,9 @@ import {
     isFunction,
     isNil,
     isString,
-    merge,
     max,
-    maxBy
+    maxBy,
+    merge
 } from 'lodash';
 import PT from 'prop-types';
 import {createRef, isValidElement} from 'react';
@@ -60,6 +62,12 @@ export const [Grid, grid] = hoistCmp.withFactory({
     model: uses(GridModel),
     className: 'xh-grid',
 
+    /**
+     * @param {GridModel} model
+     * @param {string} className
+     * @param props
+     * @param ref
+     */
     render({model, className, ...props}, ref) {
         apiRemoved(props.hideHeaders, 'hideHeaders', 'Specify hideHeaders on the GridModel instead.');
         apiRemoved(props.onKeyDown, 'onKeyDown', 'Specify onKeyDown on the GridModel instead.');
@@ -68,29 +76,30 @@ export const [Grid, grid] = hoistCmp.withFactory({
         apiRemoved(props.onCellClicked, 'onCellClicked', 'Specify onCellClicked on the GridModel instead.');
         apiRemoved(props.onCellDoubleClicked, 'onCellDoubleClicked', 'Specify onCellDoubleClicked on the GridModel instead.');
 
-        const impl = useLocalModel(() => new GridLocalModel(model, props)),
-            platformColChooser = XH.isMobileApp ? mobileColChooser : desktopColChooser;
+        const {store, treeMode, treeStyle, colChooserModel, filterModel} = model,
+            impl = useLocalModel(() => new GridLocalModel(model, props)),
+            platformColChooser = XH.isMobileApp ? mobileColChooser : desktopColChooser,
+            maxDepth = impl.isHierarchical ? store.maxDepth : null;
 
-        // Don't render the agGridReact element with data or columns. Instead rely on API methods
+        className = classNames(
+            className,
+            impl.isHierarchical ? `xh-grid--hierarchical xh-grid--max-depth-${maxDepth}` : 'xh-grid--flat',
+            treeMode ? getTreeStyleClasses(treeStyle) : null
+        );
+
         return fragment(
             frame({
-                className: classNames(
-                    className,
-                    impl.isHierarchical ? 'xh-grid--hierarchical' : 'xh-grid--flat',
-                    model.treeMode ? getTreeStyleClasses(model.treeStyle) : null
-                ),
-                item: agGrid({
-                    ...getLayoutProps(props),
-                    ...impl.agOptions
-                }),
+                className,
+                item: agGrid({...getLayoutProps(props), ...impl.agOptions}),
                 onKeyDown: impl.onKeyDown,
                 ref: composeRefs(impl.viewRef, ref)
             }),
-            (model.colChooserModel ? platformColChooser({model: model.colChooserModel}) : null),
-            (model.filterModel ? gridFilterDialog({model: model.filterModel}) : null)
+            (colChooserModel ? platformColChooser({model: colChooserModel}) : null),
+            (filterModel ? gridFilterDialog({model: filterModel}) : null)
         );
     }
 });
+
 Grid.MULTIFIELD_ROW_HEIGHT = 38;
 
 Grid.propTypes = {
@@ -98,7 +107,7 @@ Grid.propTypes = {
      * Options for ag-Grid's API.
      *
      * This constitutes an 'escape hatch' for applications that need to get to the underlying
-     * ag-Grid API.  It should be used with care. Settings made here might be overwritten and/or
+     * ag-Grid API. It should be used with care. Settings made here might be overwritten and/or
      * interfere with the implementation of this component and its use of the ag-Grid API.
      *
      * Note that changes to these options after the component's initial render will be ignored.
@@ -121,9 +130,13 @@ Grid.propTypes = {
 //------------------------
 class GridLocalModel extends HoistModel {
 
+    /** @member {GridModel} */
     model;
+    /** @member {Object} */
     agOptions;
+    /** @member {RefObject} */
     viewRef = createRef();
+    /** @member {number} */
     fixedRowHeight;
 
     getRowHeight(node) {
@@ -137,7 +150,7 @@ class GridLocalModel extends HoistModel {
         ]);
     }
 
-    // Do any root level records have children?
+    /** @returns {boolean} - true if any root-level records have children */
     @computed
     get isHierarchical() {
         const {model} = this;
@@ -163,10 +176,10 @@ class GridLocalModel extends HoistModel {
         this.addReaction(this.rowHeightReaction());
         this.addReaction(this.validationDisplayReaction());
 
-        this.agOptions = merge(this.createDefaultAgOptions(props), props.agOptions || {});
+        this.agOptions = merge(this.createDefaultAgOptions(), props.agOptions || {});
     }
 
-    createDefaultAgOptions(props) {
+    createDefaultAgOptions() {
         const {model} = this;
 
         // 'immutableData' and 'rowDataChangeDetectionStrategy' props both deal with a *new* sets of rowData.
@@ -199,10 +212,10 @@ class GridLocalModel extends HoistModel {
             getRowHeight: ({node}) => this.getRowHeight(node),
             getRowClass: ({data}) => model.rowClassFn ? model.rowClassFn(data) : null,
             noRowsOverlayComponentFramework: observer(() => div(this.emptyText)),
-            onRowDoubleClicked: model.onRowDoubleClicked,
             onCellClicked: model.onCellClicked,
             onCellDoubleClicked: model.onCellDoubleClicked,
             onRowClicked: this.onRowClicked,
+            onRowDoubleClicked: model.onRowDoubleClicked,
             onRowGroupOpened: this.onRowGroupOpened,
             onSelectionChanged: this.onSelectionChanged,
             onDragStopped: this.onDragStopped,
@@ -210,6 +223,8 @@ class GridLocalModel extends HoistModel {
             onColumnRowGroupChanged: this.onColumnRowGroupChanged,
             onColumnPinned: this.onColumnPinned,
             onColumnVisible: this.onColumnVisible,
+            onCellEditingStarted: model.onCellEditingStarted,
+            onCellEditingStopped: model.onCellEditingStopped,
             processCellForClipboard: this.processCellForClipboard,
             defaultGroupSortComparator: model.groupSortFn ? this.groupSortComparator : undefined,
             groupDefaultExpanded: 1,
@@ -374,7 +389,6 @@ class GridLocalModel extends HoistModel {
 
     sortReaction() {
         const {model} = this;
-
         return {
             track: () => [model.agColumnApi, model.sortBy],
             run: ([colApi, sortBy]) => {